refactor: move real value conversion functions to helper module
